+++ Packet++/src/SipLayer.cpp
@@ -801,7 +801,7 @@ void SipResponseFirstLine::setVersion(std::string newVersion)
 
 	if (newVersion.length() != m_Version.length())
 	{
-		LOG_ERROR("Expected version length is %d characters in the format of SIP/x.y", m_Version.length());
+		LOG_ERROR("Expected version length is %d characters in the format of SIP/x.y", (int)m_Version.length());
 		return;
 	}
